python/requirements.txt
scikit-learn
numpy
scipy
six
Pillow
tqdm
